+++ packages/react/__tests__/exports.ts
@@ -1406,6 +1406,7 @@ describe('@aws-amplify/ui-react', () => {
           "primitives",
           "useAmplify",
           "useAuthenticator",
+          "useBreakpointValue",
           "usePagination",
           "useTheme",
           "withAuthenticator",